Improve memory usage
Until now, two groups could be committed to memory at a time. With a
large group size of 100,000 rows, it could require considerable memory
in the LOH (Large Object Heap). Now only the header and two rows
(previous and current) are kept in memory at a time, thus reducing
memory usage drastically and irrespective of the group/chunk size.